Regularly evaluate feature flags
