+++ rust/telemetry/Cargo.toml
@@ -7,7 +7,7 @@ license = { workspace = true }
 [dependencies]
 anyhow = { workspace = true }
 parking_lot = { workspace = true }
-reqwest = { workspace = true, features = ["blocking"] }
+reqwest = { workspace = true }
 sentry = { workspace = true, features = ["contexts", "backtrace", "debug-images", "panic", "reqwest", "rustls", "tracing"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
